+++ apps/dine-pleiepenger/src/components/innsendt-søknad-liste/InnsendtSøknadListe.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react';
 import mockSøknader from '../../../api-mock-server/mockdata/soknader.json';
 import { withEmptyPage } from '../../storybook/hooks/withEmptyPage';
 import { withIntl } from '../../storybook/hooks/withIntl';
-import { InnsendtSøknad } from '../../types/Søknad';
+import { InnsendtSøknad } from '../../types/InnsendtSøknad';
 import InnsendtSøknadListe from './InnsendtSøknadListe';
 
 const meta: Meta<typeof InnsendtSøknadListe> = {