+++ apps/dine-pleiepenger/src/pages/api/appStatus.api.ts
@@ -1,11 +1,10 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { createChildLogger } from '@navikt/next-logger';
-import { withAuthenticatedApi } from '../../auth/withAuthentication';
-import { getXRequestId } from '../../utils/apiUtils';
-import { SanityConfig, fetchStatus } from '@navikt/appstatus-react-ds';
-import { browserEnv } from '../../utils/env';
-import { APPLICATION_KEY } from '../_app.page';
+import { fetchStatus, SanityConfig } from '@navikt/appstatus-react-ds';
 import { ApplicationState } from '@navikt/appstatus-react-ds/src/hooks/useGetApplicationStatus';
+import { withAuthenticatedApi } from '../../auth/withAuthentication';
+import { browserEnv } from '../../utils/env';
+import { getLogger } from '../../utils/getLogCorrelationID';
+import { AMPLITUDE_APPLICATION_KEY } from '../_app.page';
 
 const sanityConfig: SanityConfig = {
     projectId: browserEnv.NEXT_PUBLIC_APPSTATUS_PROJECT_ID,
@@ -13,17 +12,16 @@ const sanityConfig: SanityConfig = {
 };
 
 export const fetchAppStatus = async (): Promise<ApplicationState | undefined> => {
-    return await fetchStatus(APPLICATION_KEY, sanityConfig);
+    return await fetchStatus(AMPLITUDE_APPLICATION_KEY, sanityConfig);
 };
 
 async function handler(req: NextApiRequest, res: NextApiResponse) {
-    const childLogger = createChildLogger(getXRequestId(req));
-    childLogger.info(`Henter appStatus`);
+    const logger = getLogger(req);
+    logger.info(`Henter appStatus`);
     try {
         res.send(await fetchAppStatus());
     } catch (err) {
-        const childLogger = createChildLogger(getXRequestId(req));
-        childLogger.error(`Hent appStatus feilet: ${err}`);
+        logger.error(`Hent appStatus feilet: ${err}`);
         res.status(500).json({ error: 'Kunne ikke hente appStatus' });
     }
 }