+++ apps/dine-pleiepenger/src/pages/api/saker.api.ts
@@ -1,10 +1,9 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { createChildLogger } from '@navikt/next-logger';
-import { withAuthenticatedApi } from '../../auth/withAuthentication';
-import { fetchSaker } from '../../server/apiService';
-import { getXRequestId } from '../../utils/apiUtils';
-import { PleietrengendeMedSak } from '../../server/api-models/PleietrengendeMedSakSchema';
 import axios from 'axios';
+import { withAuthenticatedApi } from '../../auth/withAuthentication';
+import { PleietrengendeMedSak } from '../../server/api-models/PleietrengendeMedSakSchema';
+import { fetchSaker } from '../../server/apiService';
+import { getLogger } from '../../utils/getLogCorrelationID';
 
 export const sakerFetcher = async (url: string): Promise<PleietrengendeMedSak[]> =>
     axios.get(url).then((res) => res.data);
@@ -14,8 +13,7 @@ async function handler(req: NextApiRequest, res: NextApiResponse) {
         const data = await fetchSaker(req);
         res.send(data);
     } catch (err) {
-        const childLogger = createChildLogger(getXRequestId(req));
-        childLogger.error(`Hent saker feilet: ${err}`);
+        getLogger(req).error(`Hent saker feilet: ${err}`);
         res.status(500).json({ error: 'Kunne ikke hente saker' });
     }
 }