+++ apps/dine-pleiepenger/src/server/api-models/BehandlingSchema.ts
@@ -1,14 +1,14 @@
 import { z } from 'zod';
-import { AksjonspunktSchema } from './AksjonspunktSchema';
-import { SøknadSchema } from './SøknadSchema';
-import { Behandlingsstatus } from './Behandlingsstatus';
 import { parseMaybeDateStringToDate } from '../../utils/jsonParseUtils';
+import { AksjonspunktSchema } from './AksjonspunktSchema';
+import { Behandlingsstatus } from './Behandlingsstatus';
+import { InnsendelseSchema } from './InnsendelseSchema';
 
 export type Behandling = z.infer<typeof BehandlingSchema>;
 
 export const BehandlingSchema = z.object({
     status: z.nativeEnum(Behandlingsstatus),
-    søknader: z.array(SøknadSchema).nullable(),
+    innsendelser: z.array(InnsendelseSchema),
     opprettetTidspunkt: z.preprocess(parseMaybeDateStringToDate, z.date()).or(z.undefined()).or(z.null()),
     avsluttetTidspunkt: z.preprocess(parseMaybeDateStringToDate, z.date()).or(z.undefined()).or(z.null()),
     aksjonspunkter: z.array(AksjonspunktSchema),