+++ apps/dine-pleiepenger/src/types/EttersendelseType.ts
@@ -1,0 +1,4 @@
+export enum Ettersendelsestype {
+    legeerklæring = 'LEGEERKLÆRING',
+    annet = 'ANNET',
+}